requirements.txt
pyTelegramBotAPI==4.14.0
requests==2.31.0
python-dotenv==1.0.0
yt-dlp==2023.12.30
mutagen==1.47.0
Pillow==10.1.0
beautifulsoup4==4.12.2
lxml==4.9.4